(trunk)Fix compile errors on Ubuntu Quantal.
git-svn-id: http://svn.gplates.org/projects/gplates/trunk@13567 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/opengl/GLProjectionUtils.cc
@@ -180,15 +180,15 @@ GPlatesOpenGL::GLProjectionUtils::get_min_pixel_size_on_unit_sphere(
 
 	const double window_xy_coords[9][2] =
 	{
-		{viewport.x(),                          viewport.y()                           },
-		{viewport.x() + 0.5 * viewport.width(), viewport.y()                           },
-		{viewport.x() + viewport.width(),       viewport.y()                           },
-		{viewport.x(),                          viewport.y() + 0.5 * viewport.height() },
-		{viewport.x() + 0.5 * viewport.width(), viewport.y() + 0.5 * viewport.height() },
-		{viewport.x() + viewport.width(),       viewport.y() + 0.5 * viewport.height() },
-		{viewport.x(),                          viewport.y() + viewport.height()       },
-		{viewport.x() + 0.5 * viewport.width(), viewport.y() + viewport.height()       },
-		{viewport.x() + viewport.width(),       viewport.y() + viewport.height()       }
+		{double(viewport.x()),                          double(viewport.y())                           },
+                {double(viewport.x() + 0.5 * viewport.width()), double(viewport.y())                           },
+                {double(viewport.x() + viewport.width()),       double(viewport.y())                           },
+                {double(viewport.x()),                          double(viewport.y() + 0.5 * viewport.height()) },
+                {double(viewport.x() + 0.5 * viewport.width()), double(viewport.y() + 0.5 * viewport.height()) },
+                {double(viewport.x() + viewport.width()),       double(viewport.y() + 0.5 * viewport.height()) },
+                {double(viewport.x()),                          double(viewport.y() + viewport.height())       },
+                {double(viewport.x() + 0.5 * viewport.width()), double(viewport.y() + viewport.height())       },
+                {double(viewport.x() + viewport.width()),       double(viewport.y() + viewport.height())       }
 	};
 
 	// Iterate over all sample points and project onto the unit sphere in model space.
